move data to another file
