+++ lib/deprecated-services.js
@@ -1,0 +1,14 @@
+'use strict';
+
+const deprecatedServices = {
+  'gittip': new Date('2017-12-29'),
+  'gratipay': new Date('2017-12-29'),
+  'gemnasium': new Date('2018-05-15'),
+  'snap': new Date('2018-01-23'),
+  'snap-ci': new Date('2018-01-23'),
+  'cauditor': new Date('2018-02-15'),
+};
+
+module.exports = {
+  deprecatedServices
+}
