+++ lib/deprecation-helpers.js
@@ -1,15 +1,7 @@
 'use strict';
 
 const { makeBadgeData, setBadgeColor } = require('./badge-data');
-
-const deprecatedServices = {
-  'gittip': new Date('2017-12-29'),
-  'gratipay': new Date('2017-12-29'),
-  'gemnasium': new Date('2018-05-15'),
-  'snap': new Date('2018-01-23'),
-  'snap-ci': new Date('2018-01-23'),
-  'cauditor': new Date('2018-02-15'),
-};
+const { deprecatedServices } = require('./deprecated-services');
 
 const isDeprecated = function (service, now=new Date(), depServices=deprecatedServices) {
   if (!(service in depServices)) {
